Remove unneeded code from tests/t09-null-prereq
The variable $prereq2 only ever equals the empty string.  This commit
replaces the variable with the literal empty string.  This allows for
the removal of the variable's declaration and the renaming of the
variable $prereq1 to $prereq.

This commit also removes the call to exit.  Removing this line does not
change the behavior of the test in any way.